code/requirements.txt
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
